test for tax lot property export
