+++ seed/models/cycles.py
@@ -24,3 +24,4 @@ class Cycle(models.Model):
     class Meta:
         ordering = ['-created']
         get_latest_by = 'created'
+
